Add popover capture picker using melt ui
