+++ frontend/package.json
@@ -19,6 +19,8 @@
 	"devDependencies": {
 		"@floating-ui/core": "^1.3.1",
 		"@hey-api/openapi-ts": "^0.43.0",
+		"@melt-ui/pp": "^0.3.2",
+		"@melt-ui/svelte": "^0.86.2",
 		"@playwright/test": "^1.34.3",
 		"@rgossiaux/svelte-headlessui": "^2.0.0",
 		"@sveltejs/adapter-static": "^3.0.0",
@@ -128,8 +130,8 @@
 		"ol": "^7.4.0",
 		"openai": "^4.57.2",
 		"p-limit": "^6.1.0",
-		"pdfjs-dist": "4.8.69",
 		"panzoom": "^9.4.3",
+		"pdfjs-dist": "4.8.69",
 		"quill": "^1.3.7",
 		"rfc4648": "^1.5.3",
 		"svelte-carousel": "^1.0.25",
