+++ frontend/svelte.config.js
@@ -1,15 +1,18 @@
 import preprocess from 'svelte-preprocess'
 import adapter from '@sveltejs/adapter-static'
+import { preprocessMeltUI, sequence } from '@melt-ui/pp'
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	// Consult https://github.com/sveltejs/svelte-preprocess
 	// for more information about preprocessors
-	preprocess: [
+	preprocess: sequence([
 		preprocess({
 			postcss: true
-		})
-	],
+		}),
+			preprocessMeltUI()
+		]),
+
 	kit: {
 		adapter:
 			process.env.CLOUDFLARE || process.env.NOCATCHALL
